Moved snapshot to its own file. Update how snapshot gets a frame.
